+++ MoltenVK/MoltenVK/Commands/MVKCommandBuffer.h
@@ -52,6 +52,16 @@ typedef uint64_t MVKMTLCommandBufferID;
 typedef struct MVKCommandEncodingContext {
 	NSUInteger mtlVisibilityResultOffset = 0;
 	const MVKMTLBufferAllocation* visibilityResultBuffer = nullptr;
+
+	MVKRenderPass* getRenderPass() { return _renderPass; }
+	MVKFramebuffer* getFramebuffer() { return _framebuffer; }
+	void setRenderingContext(MVKRenderPass* renderPass, MVKFramebuffer* framebuffer);
+	VkRenderingFlags getRenderingFlags() { return _renderPass ? _renderPass->getRenderingFlags() : 0; }
+	~MVKCommandEncodingContext();
+
+private:
+	MVKRenderPass* _renderPass = nullptr;
+	MVKFramebuffer* _framebuffer = nullptr;
 } MVKCommandEncodingContext;
 
 
@@ -172,13 +182,15 @@ protected:
 
 	MVKCommand* _head = nullptr;
 	MVKCommand* _tail = nullptr;
-	uint32_t _commandCount;
+	MVKSmallVector<VkFormat, kMVKDefaultAttachmentCount> _colorAttachmentFormats;
 	MVKCommandPool* _commandPool;
-	std::atomic_flag _isExecutingNonConcurrently;
 	VkCommandBufferInheritanceInfo _secondaryInheritanceInfo;
+	VkCommandBufferInheritanceRenderingInfo _inerhitanceRenderingInfo;
 	id<MTLCommandBuffer> _prefilledMTLCmdBuffer = nil;
     MVKCommandEncodingContext* _immediateCmdEncodingContext = nullptr;
     MVKCommandEncoder* _immediateCmdEncoder = nullptr;
+	uint32_t _commandCount;
+	std::atomic_flag _isExecutingNonConcurrently;
 	bool _isSecondary;
 	bool _doesContinueRenderPass;
 	bool _canAcceptCommands;
@@ -223,7 +235,7 @@ public:
 						 VkSubpassContents subpassContents,
 						 MVKRenderPass* renderPass,
 						 MVKFramebuffer* framebuffer,
-						 VkRect2D& renderArea,
+						 const VkRect2D& renderArea,
 						 MVKArrayRef<VkClearValue> clearValues,
 						 MVKArrayRef<MVKImageView*> attachments,
 						 MVKArrayRef<MVKArrayRef<MTLSamplePosition>> subpassSamplePositions);
@@ -237,6 +249,9 @@ public:
 	/** Sets the dynamic custom sample positions to use when rendering. */
 	void setDynamicSamplePositions(MVKArrayRef<MTLSamplePosition> dynamicSamplePositions);
 
+	/** Begins dynamic rendering. */
+	void beginRendering(MVKCommand* rendCmd, const VkRenderingInfo* pRenderingInfo);
+
 	/** Begins a Metal render pass for the current render subpass. */
 	void beginMetalRenderPass(MVKCommandUse cmdUse);
 
@@ -244,7 +259,7 @@ public:
 	void encodeStoreActions(bool storeOverride = false);
 
 	/** Returns whether or not we are presently in a render pass. */
-	bool isInRenderPass() { return _renderPass != nullptr; }
+	bool isInRenderPass() { return _pEncodingContext->getRenderPass() != nullptr; }
 
 	/** Returns the render subpass that is currently active. */
 	MVKRenderSubpass* getSubpass();
@@ -292,6 +307,9 @@ public:
 
 	/** Ends the current renderpass. */
 	void endRenderpass();
+
+	/** Ends the current dymamic rendering. */
+	void endRendering();
 
 	/** 
 	 * Ends all encoding operations on the current Metal command encoder.
@@ -453,8 +471,6 @@ protected:
 	} GPUCounterQuery;
 
 	VkSubpassContents _subpassContents;
-	MVKRenderPass* _renderPass;
-	MVKFramebuffer* _framebuffer;
 	MVKCommand* _lastMultiviewPassCmd;
 	uint32_t _renderSubpassIndex;
 	uint32_t _multiviewPassIndex;