+++ MoltenVK/MoltenVK/GPUObjects/MVKFramebuffer.h
@@ -70,3 +70,11 @@ protected:
 	uint32_t _layerCount;
 };
 
+
+#pragma mark -
+#pragma mark Support functions
+
+/** Returns an image-less MVKFramebuffer object created from the rendering info. */
+MVKFramebuffer* mvkCreateFramebuffer(MVKDevice* device,
+									 const VkRenderingInfo* pRenderingInfo,
+									 MVKRenderPass* mvkRenderPass);