+++ Docs/Whats_New.md
@@ -19,11 +19,13 @@ MoltenVK 1.1.10
 Released TBD
 
 - Add support for extensions:
+	- `VK_KHR_dynamic_rendering`
 	- `VK_KHR_portability_enumeration` support added to `MoltenVK_icd.json`, and documentation
 	  updated to indicate the impact of the `VK_KHR_portability_enumeration` extension during 
 	  runtime loading on *macOS* via the *Vulkan Loader*.
 - Fix error where previously bound push constants can override a descriptor buffer binding 
   used by a subsequent pipeline that does not use push constants.
+- Support attachment clearing when some clearing formats are not specified.